commit 13-10-25 v2

+++ src/main/java/com/impresiones/service/FuncionarioService.java
@@ -6,5 +6,5 @@ import java.util.List;
 public interface FuncionarioService {
     List<Funcionario> listarFuncionarios();
     void guardar(Funcionario funcionario);
-    void eliminar(Long id);
+    void eliminar(int id);
 }
